requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
deap==1.3.3
debugpy==1.6.7
decision-tree-id3==0.1.2
decorator==5.1.1
executing==1.2.0
fonttools==4.39.3
graphviz==0.20.1
idna==3.4
importlib-metadata==6.4.1
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
nose==1.3.7
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
regex==2023.5.5
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn-deap==0.3.0
sklearn-genetic-opt==0.10.1
stack-data==0.6.2
textblob==0.17.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vader-sentiment==3.2.1.1
vaderSentiment==3.3.2
wcwidth==0.2.6
xgboost==1.7.5
zipp==3.15.0